requirements.txt
playwright~=1.51.0
pytest~=8.3.5
pillow~=11.1.0